refactor(clients): rename src/pocketpaw/integrations/ → src/pocketpaw/clients/
The previous "integrations" name was vague — it could mean "external
service integrations" or "API integrations" or "ee/cloud integrations."
"clients" is the actual role: HTTP / SDK clients for third-party
services (Gmail, Google Calendar, Google Docs, Google Drive, Reddit,
Spotify) plus shared OAuth + token storage.

Rename only — zero behaviour change. Every import site rewritten via
sed; doc references updated; ruff clean.

Layer responsibilities (now explicit):
- src/pocketpaw/clients/      HTTP / SDK clients (low-level: tokens,
                              MIME, base64, HTTP calls). Stateful.
- src/pocketpaw/connectors/   Connector protocol + adapters wrapping
                              clients. Stateless.
- src/pocketpaw/tools/builtin/  Agent-facing tools. Hand-tuned LLM
                              response formatting.
- ee/cloud/connectors/        Tenanted state + REST router (only
                              enterprise piece).

Tests
- 199 connector + integration tests pass (199 / 199, no regressions)
- ruff check clean on src/pocketpaw/, ee/cloud/connectors/, tests/connectors/

What's NOT in this PR
- pocketpaw/api/v1/oauth_integrations.py kept as-is — different concept
  (REST endpoint for OAuth flows, not a service client). Could rename
  later but not load-bearing.
