+++ src/pocketpaw/dashboard.py
@@ -766,7 +766,7 @@ async def oauth_authorize(service: str = Query("google_gmail")):
                 detail="Google OAuth Client ID not configured. Set it in Settings first.",
             )
 
-    from pocketpaw.integrations.oauth import OAuthManager
+    from pocketpaw.clients.oauth import OAuthManager
 
     manager = OAuthManager()
     redirect_uri = f"http://localhost:{settings.web_port}/oauth/callback"
@@ -798,8 +798,8 @@ async def oauth_callback(
         return HTMLResponse("<h2>Missing authorization code</h2>")
 
     try:
-        from pocketpaw.integrations.oauth import OAuthManager
-        from pocketpaw.integrations.token_store import TokenStore
+        from pocketpaw.clients.oauth import OAuthManager
+        from pocketpaw.clients.token_store import TokenStore
 
         settings = Settings.load()
         manager = OAuthManager(TokenStore())
